Refactor swap package structure
