+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/approve/SwapApproveModule.kt
@@ -10,7 +10,7 @@ import io.horizontalsystems.bankwallet.core.ethereum.EthereumTransactionService
 import io.horizontalsystems.bankwallet.core.factories.FeeRateProviderFactory
 import io.horizontalsystems.bankwallet.core.providers.EthereumFeeRateProvider
 import io.horizontalsystems.bankwallet.entities.CoinValue
-import io.horizontalsystems.bankwallet.modules.swap_new.SwapAllowanceService
+import io.horizontalsystems.bankwallet.modules.swap_new.allowance.SwapAllowanceService
 import io.horizontalsystems.ethereumkit.models.Address
 import io.reactivex.Observable
 import java.math.BigInteger
